backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
Pillow==10.1.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-filter==23.3
drf-spectacular==0.26.5
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
factory-boy==3.3.0
pytest-django==4.7.0
coverage==7.3.2
